Move LayoutConfig.tsx to top-level /storyblok directory
